+++ cmd/nodecmd/helpers.go
@@ -6,13 +6,18 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"strings"
 	"sync"
 
 	"github.com/ava-labs/avalanche-cli/pkg/ansible"
+	"github.com/ava-labs/avalanche-cli/pkg/constants"
 	"github.com/ava-labs/avalanche-cli/pkg/models"
 	"github.com/ava-labs/avalanche-cli/pkg/ssh"
 	"github.com/ava-labs/avalanche-cli/pkg/utils"
 	"github.com/ava-labs/avalanche-cli/pkg/ux"
+	"github.com/ava-labs/avalanche-cli/pkg/vm"
+
+	"golang.org/x/exp/slices"
 )
 
 func checkHostsAreHealthy(hosts []*models.Host) ([]string, error) {
@@ -185,6 +190,136 @@ func parseBootstrappedOutput(byteValue []byte) (bool, error) {
 		}
 	}
 	return false, errors.New("unable to parse node bootstrap status")
+}
+
+func checkAvalancheGoVersionCompatibleAtHosts(hosts []*models.Host, subnetName string) ([]string, error) {
+	ux.Logger.PrintToUser("Checking compatibility of node(s) avalanche go version with Subnet EVM RPC of subnet %s ...", subnetName)
+	compatibleVersions := []string{}
+	incompatibleNodes := []string{}
+	wg := sync.WaitGroup{}
+	wgResults := models.NodeResults{}
+	for _, host := range hosts {
+		wg.Add(1)
+		go func(nodeResults *models.NodeResults, host *models.Host) {
+			defer wg.Done()
+			if resp, err := ssh.RunSSHCheckAvalancheGoVersion(host); err != nil {
+				nodeResults.AddResult(host.NodeID, nil, err)
+				return
+			} else {
+				if avalancheGoVersion, err := parseAvalancheGoOutput(resp); err != nil {
+					nodeResults.AddResult(host.NodeID, nil, err)
+				} else {
+					nodeResults.AddResult(host.NodeID, avalancheGoVersion, err)
+				}
+			}
+		}(&wgResults, host)
+	}
+	wg.Wait()
+	if wgResults.HasErrors() {
+		return nil, fmt.Errorf("failed to get avalanchego version for node(s) %s", wgResults.GetErrorHostMap())
+	}
+	for nodeID, avalancheGoVersion := range wgResults.GetResultMap() {
+		sc, err := app.LoadSidecar(subnetName)
+		if err != nil {
+			return nil, err
+		}
+		compatibleVersions, err = checkForCompatibleAvagoVersion(sc.RPCVersion)
+		if err != nil {
+			return nil, err
+		}
+		if !slices.Contains(compatibleVersions, fmt.Sprintf("%v", avalancheGoVersion)) {
+			incompatibleNodes = append(incompatibleNodes, nodeID)
+		}
+	}
+	if len(incompatibleNodes) > 0 {
+		ux.Logger.PrintToUser(fmt.Sprintf("Compatible Avalanche Go versions are %s", strings.Join(compatibleVersions, ", ")))
+	}
+	return incompatibleNodes, nil
+}
+
+func checkAvalancheGoVersionCompatible(clusterName, subnetName string) ([]string, error) {
+	ux.Logger.PrintToUser(fmt.Sprintf("Checking compatibility of avalanche go version in cluster %s with Subnet EVM RPC of subnet %s ...", clusterName, subnetName))
+	compatibleVersions := []string{}
+	incompatibleNodes := []string{}
+	hosts, err := ansible.GetInventoryFromAnsibleInventoryFile(app.GetAnsibleInventoryDirPath(clusterName))
+	if err != nil {
+		return nil, err
+	}
+	wg := sync.WaitGroup{}
+	wgResults := models.NodeResults{}
+	for _, host := range hosts {
+		wg.Add(1)
+		go func(nodeResults *models.NodeResults, host *models.Host) {
+			defer wg.Done()
+			if err := host.Connect(); err != nil {
+				nodeResults.AddResult(host.NodeID, nil, err)
+				return
+			}
+			defer func() {
+				if err := host.Disconnect(); err != nil {
+					nodeResults.AddResult(host.NodeID, nil, err)
+				}
+			}()
+			if resp, err := ssh.RunSSHCheckAvalancheGoVersion(host); err != nil {
+				nodeResults.AddResult(host.NodeID, nil, err)
+				return
+			} else {
+				if avalancheGoVersion, err := parseAvalancheGoOutput(resp); err != nil {
+					nodeResults.AddResult(host.NodeID, nil, err)
+				} else {
+					nodeResults.AddResult(host.NodeID, avalancheGoVersion, err)
+				}
+			}
+		}(&wgResults, host)
+	}
+	wg.Wait()
+	if wgResults.HasErrors() {
+		return nil, fmt.Errorf("failed to get avalanchego version for node(s) %s", wgResults.GetErrorHostMap())
+	}
+	for nodeID, avalancheGoVersion := range wgResults.GetResultMap() {
+		sc, err := app.LoadSidecar(subnetName)
+		if err != nil {
+			return nil, err
+		}
+		compatibleVersions, err = checkForCompatibleAvagoVersion(sc.RPCVersion)
+		if err != nil {
+			return nil, err
+		}
+		if !slices.Contains(compatibleVersions, fmt.Sprintf("%v", avalancheGoVersion)) {
+			incompatibleNodes = append(incompatibleNodes, nodeID)
+		}
+	}
+	if len(incompatibleNodes) > 0 {
+		ux.Logger.PrintToUser(fmt.Sprintf("Compatible Avalanche Go versions are %s", strings.Join(compatibleVersions, ", ")))
+	}
+	return incompatibleNodes, nil
+}
+
+func parseAvalancheGoOutput(byteValue []byte) (string, error) {
+	var result map[string]interface{}
+	if err := json.Unmarshal(byteValue, &result); err != nil {
+		return "", err
+	}
+	nodeIDInterface, ok := result["result"].(map[string]interface{})
+	if ok {
+		vmVersions, ok := nodeIDInterface["vmVersions"].(map[string]interface{})
+		if ok {
+			avalancheGoVersion, ok := vmVersions["platform"].(string)
+			if ok {
+				return avalancheGoVersion, nil
+			}
+		}
+	}
+	return "", nil
+}
+
+func checkForCompatibleAvagoVersion(configuredRPCVersion int) ([]string, error) {
+	compatibleAvagoVersions, err := vm.GetAvailableAvalancheGoVersions(
+		app, configuredRPCVersion, constants.AvalancheGoCompatibilityURL)
+	if err != nil {
+		return nil, err
+	}
+	return compatibleAvagoVersions, nil
 }
 
 func connectHosts(hosts []*models.Host) models.NodeResults {
